+++ Wrappers/Python/cil/optimisation/operators/GradientOperator.py
@@ -39,6 +39,9 @@ class GradientOperator(LinearOperator):
     :type gm_domain: `ImageGeometry`
     :param bnd_cond: Set the boundary conditions to use 'Neumann' or 'Periodic', defaults to 'Neumann'
     :type bnd_cond: str, optional    
+    :return: returns a BlockDataContainer containing images of the derivatives order given by `dimension_labels`\
+        i.e. ['horizontal_y','horizontal_x'] will return [d('horizontal_y'), d('horizontal_x')]
+    :rtype: BlockDataContainer
     
     :param \**kwargs:
         See below
@@ -50,8 +53,9 @@ class GradientOperator(LinearOperator):
           'c' or 'numpy', defaults to 'c' if correlation is 'SpaceChannels' or channels = 1
         * *num_threads* (``int``) --
           If backend is 'c' specify the number of threads to use. Default is number of cpus/2          
-                 
-                 
+        * *split* (``boolean``) --
+          If 'True', and backend 'C' will return a BlockDataContainer with grouped spatial domains. i.e. [Channel, [Z, Y, X]], otherwise [Channel, Z, Y, X]
+                               
         Example (2D): 
         .. math::
         
@@ -140,79 +144,48 @@ class Gradient_numpy(LinearOperator):
         # Call FiniteDiff operator 
         self.method = method
         self.FD = FiniteDifferenceOperator(domain_geometry, direction = 0, method = self.method, bnd_cond = self.bnd_cond)
-                
         
-        if self.correlation==CORRELATION_SPACE:
-            
-            if domain_geometry.channels > 1:
-                
-                range_geometry = BlockGeometry(*[domain_geometry for _ in range(domain_geometry.length-1)] )
+        self.ndim = len(domain_geometry.shape)
+        self.ind = list(range(self.ndim))
 
-                if self.size_dom_gm == 4:
-                    # 3D + Channel
-                    expected_order = [ImageGeometry.CHANNEL, ImageGeometry.VERTICAL, ImageGeometry.HORIZONTAL_Y, ImageGeometry.HORIZONTAL_X]
+        if self.correlation==CORRELATION_SPACE and 'channel' in domain_geometry.dimension_labels:
+            self.ndim -= 1
+            self.ind.remove(domain_geometry.dimension_labels.index('channel'))
 
-                else:
-                    # 2D + Channel
-                    expected_order = [ImageGeometry.CHANNEL, ImageGeometry.HORIZONTAL_Y, ImageGeometry.HORIZONTAL_X]
+        range_geometry = BlockGeometry(*[domain_geometry for _ in range(self.ndim) ] )
 
-                order = domain_geometry.get_order_by_label(domain_geometry.dimension_labels, expected_order)
-                
-                self.ind = order[1:]
-                
-            else:
-                # no channel info
-                range_geometry = BlockGeometry(*[domain_geometry for _ in range(domain_geometry.length) ] )
-                if self.size_dom_gm == 3:
-                    # 3D
-                    expected_order = [ImageGeometry.VERTICAL, ImageGeometry.HORIZONTAL_Y, ImageGeometry.HORIZONTAL_X]
-#                    self.voxel_size_order = [domain_geometry.voxel_size_z, domain_geometry.voxel_size_y, domain_geometry.voxel_size_x ]                    
-                    
-                else:
-                    # 2D
-                    expected_order = [ImageGeometry.HORIZONTAL_Y, ImageGeometry.HORIZONTAL_X]    
+        #get voxel spacing, if not use 1s
+        try:
+            self.voxel_size_order = list(domain_geometry.spacing)
+        except:
+            self.voxel_size_order = [1]*len(domain_geometry.shape)
 
-                self.ind = domain_geometry.get_order_by_label(domain_geometry.dimension_labels, expected_order)
-                
-        elif self.correlation==CORRELATION_SPACECHANNEL:
-            
-            if domain_geometry.channels > 1:
-                range_geometry = BlockGeometry(*[domain_geometry for _ in range(domain_geometry.length)])
-                self.ind = range(domain_geometry.length)                
-            else:
-                raise ValueError('No channels to correlate')
-                
-        self.voxel_size_order = domain_geometry.spacing                
-         
         super(Gradient_numpy, self).__init__(domain_geometry = domain_geometry, 
                                              range_geometry = range_geometry) 
         
         print("Initialised GradientOperator with numpy backend")               
         
-    def direct(self, x, out=None):        
-                
-         if out is not None:
-            
-             for i in range(self.range_geometry().shape[0]):
-                 self.FD.direction = self.ind[i]
-                 self.FD.voxel_size = self.voxel_size_order[i]
+    def direct(self, x, out=None):              
+         if out is not None:  
+             for i, axis_index in enumerate(self.ind):
+                 self.FD.direction = axis_index
+                 self.FD.voxel_size = self.voxel_size_order[axis_index]
                  self.FD.direct(x, out = out[i])
          else:
              tmp = self.range_geometry().allocate()        
-             for i in range(tmp.shape[0]):
-                 self.FD.direction = self.ind[i]
-                 self.FD.voxel_size = self.voxel_size_order[i]
+             for i, axis_index in enumerate(self.ind):
+                 self.FD.direction = axis_index
+                 self.FD.voxel_size = self.voxel_size_order[axis_index]
                  tmp.get_item(i).fill(self.FD.direct(x))
              return tmp    
         
     def adjoint(self, x, out=None):
 
         if out is not None:
-
             tmp = self.domain_geometry().allocate()            
-            for i in range(x.shape[0]):
-                self.FD.direction=self.ind[i] 
-                self.FD.voxel_size = self.voxel_size_order[i]
+            for i, axis_index in enumerate(self.ind):
+                self.FD.direction = axis_index
+                self.FD.voxel_size = self.voxel_size_order[axis_index]
                 self.FD.adjoint(x.get_item(i), out = tmp)
                 if i == 0:
                     out.fill(tmp)
@@ -220,9 +193,9 @@ class Gradient_numpy(LinearOperator):
                     out += tmp
         else:            
             tmp = self.domain_geometry().allocate()
-            for i in range(x.shape[0]):
-                self.FD.direction=self.ind[i]
-                self.FD.voxel_size = self.voxel_size_order[i]
+            for i, axis_index in enumerate(self.ind):
+                self.FD.direction = axis_index
+                self.FD.voxel_size = self.voxel_size_order[axis_index]
                 tmp += self.FD.adjoint(x.get_item(i))
             return tmp    
 
@@ -291,10 +264,11 @@ class Gradient_C(LinearOperator):
     def __init__(self, gm_domain, gm_range=None, bnd_cond = NEUMANN, **kwargs):
 
         self.num_threads = kwargs.get('num_threads',NUM_THREADS)
-
+        self.split = kwargs.get('split',False)
         self.gm_domain = gm_domain
         self.gm_range = gm_range
-        
+        self.ndim = self.gm_domain.length
+                
         #default is 'Neumann'
         self.bnd_cond = 0
         
@@ -303,8 +277,12 @@ class Gradient_C(LinearOperator):
         
         # Domain Geometry = Range Geometry if not stated
         if self.gm_range is None:
-            self.gm_range = BlockGeometry(*[gm_domain for _ in range(len(gm_domain.shape))])
-        
+            if self.split is True and 'channel' in self.gm_domain.dimension_labels:
+                self.gm_range = BlockGeometry(gm_domain, BlockGeometry(*[gm_domain for _ in range(len(gm_domain.shape)-1)]))
+            else:
+                self.gm_range = BlockGeometry(*[gm_domain for _ in range(len(gm_domain.shape))])
+                self.split = False
+
         if len(gm_domain.shape) == 4:
             # Voxel size wrt to channel direction == 1.0
             self.fd = cilacc.fdiff4D
@@ -329,31 +307,46 @@ class Gradient_C(LinearOperator):
     def ndarray_as_c_pointer(ndx):
         return ndx.ctypes.data_as(c_float_p)
         
-    def direct(self, x, out=None):
-        ndx = np.asarray(x.as_array(), dtype=np.float32)
-        #ndx , x_p = Gradient_C.datacontainer_as_c_pointer(x)
+    def direct(self, x, out=None):    
+        ndx = np.asarray(x.as_array(), dtype=np.float32, order='C')
         x_p = Gradient_C.ndarray_as_c_pointer(ndx)
         
         return_val = False
         if out is None:
             out = self.gm_range.allocate(None)
             return_val = True
-        ndout = [el.as_array() for el in out.containers]
 
+        if self.split is False:
+            ndout = [el.as_array() for el in out.containers]
+        else:
+            ind = self.gm_domain.dimension_labels.index('channel')
+            ndout = [el.as_array() for el in out.get_item(1).containers]
+            ndout.insert(ind, out.get_item(0).as_array()) #insert channels dc at correct point for channel data
+                
         #pass list of all arguments
-        #arg1 = [Gradient_C.datacontainer_as_c_pointer(out.get_item(i))[1] for i in range(self.gm_range.shape[0])]
-        arg1 = [Gradient_C.ndarray_as_c_pointer(ndout[i]) for i in range(self.gm_range.shape[0])]
+        arg1 = [Gradient_C.ndarray_as_c_pointer(ndout[i]) for i in range(len(ndout))]
         arg2 = [el for el in x.shape]
         args = arg1 + arg2 + [self.bnd_cond, 1, self.num_threads]
         self.fd(x_p, *args)
-        
-        for i in range(len(ndout)):
-            out.get_item(i).fill(ndout[i])
 
-        if any(elem != 1.0 for elem in self.voxel_size_order):
-            out /= self.voxel_size_order
-#        out /= self.voxel_size_order
-        
+        for i, el in enumerate(self.voxel_size_order):
+            if el != 1:
+                ndout[i]/=el
+
+        #fill back out in corerct (non-traivial) order
+        if self.split is False:
+            for i in range(self.ndim):
+                out.get_item(i).fill(ndout[i])
+        else:
+            ind = self.gm_domain.dimension_labels.index('channel')
+            out.get_item(0).fill(ndout[ind])
+
+            j = 0
+            for i in range(self.ndim):
+                if i != ind:
+                    out.get_item(1).get_item(j).fill(ndout[i])
+                    j +=1
+
         if return_val is True:
             return out        
 
@@ -363,29 +356,32 @@ class Gradient_C(LinearOperator):
         if out is None:
             out = self.gm_domain.allocate(None)
             return_val = True
-        ndout = out.as_array()
-        
 
-        # ndout, out_p = Gradient_C.datacontainer_as_c_pointer(out)
+        ndout = np.asarray(out.as_array(), dtype=np.float32, order='C')          
         out_p = Gradient_C.ndarray_as_c_pointer(ndout)
         
-               
-        
-        if any(elem != 1.0 for elem in self.voxel_size_order):
-            tmp = x/self.voxel_size_order
+        if self.split is False:
+            ndx = [el.as_array() for el in x.containers]
         else:
-            tmp = x
-        ndx = [el.as_array() for el in tmp.containers]
-        
-        #arg1 = [Gradient_C.datacontainer_as_c_pointer(tmp.get_item(i))[1] for i in range(self.gm_range.shape[0])]
-        arg1 = [Gradient_C.ndarray_as_c_pointer(ndx[i]) for i in range(self.gm_range.shape[0])]
+            ind = self.gm_domain.dimension_labels.index('channel')
+            ndx = [el.as_array() for el in x.get_item(1).containers]
+            ndx.insert(ind, x.get_item(0).as_array()) 
+
+        for i, el in enumerate(self.voxel_size_order):
+            if el != 1:
+                ndx[i]/=el
+
+        arg1 = [Gradient_C.ndarray_as_c_pointer(ndx[i]) for i in range(self.ndim)]
         arg2 = [el for el in out.shape]
         args = arg1 + arg2 + [self.bnd_cond, 0, self.num_threads]
 
         self.fd(out_p, *args)
         out.fill(ndout)
 
+        #reset input data
+        for i, el in enumerate(self.voxel_size_order):
+            if el != 1:
+                ndx[i]*= el
+                
         if return_val is True:
             return out
-
-