+++ Wrappers/Python/cil/processors/__init__.py
@@ -21,5 +21,8 @@ from .Normaliser import Normaliser
 from .Slicer import Slicer
 from .Binner import Binner
 from .RingRemover import RingRemover
+from .MaskGenerator import MaskGenerator
 from .AbsorptionTransmissionConverter import AbsorptionTransmissionConverter
 from .TransmissionAbsorptionConverter import TransmissionAbsorptionConverter
+from .Masker import Masker
+